testUtils plugin (#908)
* testUtils plugin

* other platforms

* more fixes

* fixes

* fixes

* platform data abstraction

* typo

* Apple build

* android build

* uwp build

* ios build

* removed empty space
