+++ Apps/ValidationTests/iOS/LibNativeBridge.mm
@@ -6,6 +6,7 @@
 #import <Babylon/Plugins/NativeEngine.h>
 #import <Babylon/Plugins/NativeOptimizations.h>
 #import <Babylon/Plugins/NativeXr.h>
+#import <Babylon/Plugins/TestUtils.h>
 #import <Babylon/Polyfills/Canvas.h>
 #import <Babylon/Polyfills/Window.h>
 #import <Babylon/Polyfills/XMLHttpRequest.h>
@@ -13,8 +14,6 @@
 
 std::unique_ptr<Babylon::Graphics> graphics{};
 std::unique_ptr<Babylon::AppRuntime> runtime{};
-
-#import <Shared/TestUtils.h>
 
 @implementation LibNativeBridge
 
@@ -60,7 +59,7 @@ std::unique_ptr<Babylon::AppRuntime> runtime{};
         // Initialize NativeXr plugin.
         Babylon::Plugins::NativeXr::Initialize(env);
 
-        Babylon::TestUtils::CreateInstance(env, nullptr);
+        Babylon::Plugins::TestUtils::Initialize(env, nullptr);
     });
 
     Babylon::ScriptLoader loader{ *runtime };
