+++ Apps/ValidationTests/macOS/ViewController.mm
@@ -6,12 +6,12 @@
 #import <Babylon/Graphics.h>
 #import <Babylon/Plugins/NativeEngine.h>
 #import <Babylon/Plugins/NativeOptimizations.h>
+#import <Babylon/Plugins/TestUtils.h>
 #import <Babylon/Polyfills/Window.h>
 #import <Babylon/Polyfills/XMLHttpRequest.h>
 #import <Babylon/Polyfills/Canvas.h>
 #import <Babylon/ScriptLoader.h>
 #import <MetalKit/MetalKit.h>
-#import <Shared/TestUtils.h>
 
 std::unique_ptr<Babylon::Graphics> graphics{};
 std::unique_ptr<Babylon::AppRuntime> runtime{};
@@ -94,7 +94,7 @@ std::unique_ptr<Babylon::AppRuntime> runtime{};
 
         Babylon::Plugins::NativeOptimizations::Initialize(env);
 
-        Babylon::TestUtils::CreateInstance(env, engineView);
+        Babylon::Plugins::TestUtils::Initialize(env, engineView);
     });
 
     Babylon::ScriptLoader loader{ *runtime };
