+++ Plugins/TestUtils/Include/Babylon/Plugins/TestUtils.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <napi/env.h>
+#include <Babylon/Graphics.h>
+
+namespace Babylon::Plugins::TestUtils
+{
+    extern int errorCode;
+    void Initialize(Napi::Env env, WindowType nativeWindowPtr);
+}
